requirements.txt
azure-functions
python-dotenv
requests
pandas
numpy
orjson
azure-monitor-query
azure-identity
xmltodict